requirements.txt
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15

langchain==0.3.27
langchain-core==0.3.80
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.11

langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.10

pydantic==2.12.4
pydantic-core==2.41.5
typing-extensions==4.15.0

pandas==2.3.3
numpy==2.3.5

matplotlib==3.10.7
seaborn==0.13.2
openpyxl==3.1.5

python-dotenv==1.2.1
requests==2.32.5
httpx==0.28.1

fastapi==0.115.2
uvicorn==0.30.5

streamlit==1.51.0
tqdm==4.67.1
